refactor(postBoard) : post board 수정
